requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
